// DELIVERABLES/Source Code/src/app/shared/models/customer/customer-filter.model.ts
export class CustomerFilter {
    business: string = null;
    group: string = null;
    evaluate: string = null;
    fromDate: number;
    toDate: number;
    month: string = null;
    hasBirthday: boolean;
    assignedEmployeeId: string = null;
    assignTo: string = null;
    assignToGroup: string = null;
}